Create custom hook to handle route loading
