+++ pages/tournaments/[tournamentId]/targets/[id].tsx
@@ -12,6 +12,7 @@ import { Tournament } from "@prisma/client";
 import { UserProvider } from "../../../../components/UserProvider";
 import { useRouter } from "next/router";
 import LoadingSpinner from "../../../../components/Common/LoadingSpinner";
+import { useRouterLoading } from "../../../../lib/hooks";
 
 const isTournamentRunning = (startTime: Date, endTime: Date) => {
   const currentTime = new Date().getTime();
@@ -178,23 +179,13 @@ export default function Target({
   currentUser
 }): JSX.Element {
   const [user, setUser] = useState(u);
-  const [isLoading, setIsLoading] = useState(false);
   const theme = useTheme();
   const isMobileView = useMediaQuery(theme.breakpoints.down("md"));
-  const router = useRouter();
 
-  const startLoading = () => setIsLoading(true);
-  const stopLoading = () => setIsLoading(false);
-  useEffect(() => {
-    router.events.on("routeChangeStart", startLoading);
-    router.events.on("routeChangeComplete", stopLoading);
-    return () => {
-      router.events.off("routeChangeStart", startLoading);
-      router.events.off("routeChangeComplete", stopLoading);
-    };
-  }, [router]);
+  const isLoading = useRouterLoading();
 
   if (isLoading) return <LoadingSpinner />;
+
   return (
     <AuthenticationRequired>
       <UserProvider user={user}>
